requirements.txt
azure-functions
mechanize